requirements.txt
# parse

django==3.2.0
# django-heroku==0.3.1
django-extensions
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-filter==21.1
# djoser==2.1.0

# Faker==13.3.0
# model-bakery==1.4.0

waitress
jinja2
webob==1.8.7
whitenoise==6.0.0

ipython==8.0.0
psycopg2-binary==2.9.3
